Add files via upload
Hey guys, here are four 'working' files to produce plots showing the galaxies.  To run the Nbody_galaxy.c code you need to run the executable with a 0 for rk4 and a 1 for leapfrog (probably easier to run with the Makefile).  All the points will then be written to a .dat file.  Currently I have it writing to a file for every 2 timesteps (making it a total of 10 images, but we may actually want it to print for every timestep.  I'm also trying to figure out the scaling on the images, but it seems like for some reason the two galaxies are collapsing in on each other, upon contact, so we need to investigate that...
I've updated my Initial conditions file to include the more accurate scientific scaling.  To plot, just run the galaxyplot_lf.py file.